refactor: 외부 클라이언트의 구현체는 infra 패키지로 이동
