+++ src/main/java/kitchenpos/menus/tobe/infra/DefaultProfanities.java
@@ -1,5 +1,6 @@
-package kitchenpos.menus.tobe.domain.vo;
+package kitchenpos.menus.tobe.infra;
 
+import kitchenpos.menus.tobe.domain.vo.Profanities;
 import org.springframework.stereotype.Component;
 
 @Component
