+++ src/test/java/kitchenpos/products/tobe/infra/FakeProfanitiesClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.tobe.domain;
+package kitchenpos.products.tobe.infra;
